fix: rename folder, add static lifetime to &str
