+++ src/ressources/Licence.rs
@@ -1,0 +1,6 @@
+pub struct LicenceShortNames {}
+
+pub struct Licence {
+    short_name: (LicenceShortNames, &'static str),
+    detail_text: &'static str
+}
