+++ src/ressources/StandAloneParagraph.rs
@@ -1,0 +1,6 @@
+use crate::ressources::Licence::Licence;
+
+pub struct StandAloneParagraph {
+    license: Licence,
+    comment: Option<&'static str>,
+}
