+++ src/ressources/mod.rs
@@ -1,0 +1,5 @@
+pub mod Copyright;
+pub mod FileParagraph;
+pub mod HeaderParagraph;
+pub mod Licence;
+pub mod StandAloneParagraph;
