--- src/resssources/Licence.rs
@@ -1,6 +1,0 @@
-pub struct LicenceShortNames {}
-
-pub struct Licence {
-    short_name: (LicenceShortNames, &str),
-    detail_text: &str
-}
