--- src/resssources/StandAloneParagraph.rs
@@ -1,4 +1,0 @@
-pub struct StandAloneParagraph {
-    license: Licence,
-    comment: Option<&str>,
-}
